feat: Add admin login for maintenance mode (#680)
* feat: Add admin login for maintenance mode

Allows admins to log in using a local account when maintenance mode is enabled.
An admin is any user that has a role with the maintenance override permission.

* feat: Add missing translations for maintenance admin login

Adds the Dutch and Polish translations for the strings added for maintenance mode admin login.

* style: Position maintenance mode overlay fixed

Changes the positioning for the MaintenanceModeOverlay from absolute to fixed.

Before, the footer would sometimes show below the overlay if the screen was narrow enough.
Giving the overlay position fixed makes it always cover the full page.