feat: SudoSOS Wrapped 2025 (#655)
* feat: Add Wrapped table to db

Adds a new table to the database that will be used to store the precomputed data for SudoSOS Wrapped.

Rather than having it's own ID per row, the rows in the Wrapped table are mapped directly to a User. Meaning userId to mapped directly to user.id.

* feat: Add /v1/users/{id}/wrapped endpoints

Adds the GET /wrapped endpoints. This endpoint will return the wrapped data for the current user.

Adds the wrapped-controller, wrapped-service and wrapped-response.

* feat: Add update wrapped cron job

* feat: Update wrapped table in maintenance